requirements.txt
pymupdf
pytesseract
pillow
pdf2image
spacy
openai
python-dotenv
streamlit
plotly
